# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lexus GS300 Arıza Teşhis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RepairStep {
  order: number;
  description: string;
  /** @nullable */
  estimatedTime?: string | null;
  /** @nullable */
  toolRequired?: string | null;
}

export type DtcCodeSeverity = typeof DtcCodeSeverity[keyof typeof DtcCodeSeverity];


export const DtcCodeSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

/**
 * DIY difficulty level
 */
export type DtcCodeDifficulty = typeof DtcCodeDifficulty[keyof typeof DtcCodeDifficulty];


export const DtcCodeDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  professional: 'professional',
} as const;

export interface DtcCode {
  /** OBD-II code e.g. P0300 */
  code: string;
  /** Short human-readable title (Turkish) */
  title: string;
  /** Vehicle system identifier */
  system: string;
  /** Human-readable system name (Turkish) */
  systemLabel?: string;
  severity: DtcCodeSeverity;
  /** Full description of the fault (Turkish) */
  description: string;
  /** Observable symptoms (Turkish) */
  symptoms: string[];
  /** Possible causes (Turkish) */
  causes: string[];
  repairSteps: RepairStep[];
  /** Approximate repair cost range in Turkish Lira */
  estimatedCost: string;
  /** DIY difficulty level */
  difficulty: DtcCodeDifficulty;
  relatedCodes?: string[];
}

export interface SymptomInput {
  /** List of symptom IDs or free text descriptions */
  symptoms: string[];
  /**
     * Optional vehicle system to narrow search
     * @nullable
     */
  system?: string | null;
}

export type DiagnosisResultConfidence = typeof DiagnosisResultConfidence[keyof typeof DiagnosisResultConfidence];


export const DiagnosisResultConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface DiagnosisMatch {
  dtcCode: DtcCode;
  /** Match score 0-100 */
  score: number;
  matchedSymptoms: string[];
}

export interface DiagnosisResult {
  matches: DiagnosisMatch[];
  confidence: DiagnosisResultConfidence;
  note?: string;
}

export interface Symptom {
  id: string;
  label: string;
  /** @nullable */
  systemHint?: string | null;
}

export interface SymptomCategory {
  id: string;
  label: string;
  symptoms: Symptom[];
}

export interface VehicleSystem {
  id: string;
  label: string;
  icon: string;
  description?: string;
  codeCount: number;
  criticalCount: number;
}

export type StatsSummaryBySeverity = {[key: string]: number};

export type StatsSummaryBySystemItem = {
  system: string;
  label: string;
  count: number;
};

export interface StatsSummary {
  totalCodes: number;
  bySeverity: StatsSummaryBySeverity;
  bySystem: StatsSummaryBySystemItem[];
}

export type ListDtcCodesParams = {
/**
 * Search by code, title, or description
 */
search?: string;
/**
 * Filter by vehicle system (engine, transmission, abs, etc.)
 */
system?: string;
/**
 * Filter by severity level
 */
severity?: ListDtcCodesSeverity;
};

export type ListDtcCodesSeverity = typeof ListDtcCodesSeverity[keyof typeof ListDtcCodesSeverity];


export const ListDtcCodesSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;