# lib/api-zod/src/generated/types/diagnosisMatch.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lexus GS300 Arıza Teşhis API
 * OpenAPI spec version: 0.1.0
 */
import type { DtcCode } from './dtcCode';

export interface DiagnosisMatch {
  dtcCode: DtcCode;
  /** Match score 0-100 */
  score: number;
  matchedSymptoms: string[];
}