# lib/api-zod/src/generated/types/diagnosisResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lexus GS300 Arıza Teşhis API
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosisMatch } from './diagnosisMatch';
import type { DiagnosisResultConfidence } from './diagnosisResultConfidence';

export interface DiagnosisResult {
  matches: DiagnosisMatch[];
  confidence: DiagnosisResultConfidence;
  note?: string;
}