# lib/api-zod/src/generated/types/dtcCode.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lexus GS300 Arıza Teşhis API
 * OpenAPI spec version: 0.1.0
 */
import type { DtcCodeDifficulty } from './dtcCodeDifficulty';
import type { DtcCodeSeverity } from './dtcCodeSeverity';
import type { RepairStep } from './repairStep';

export interface DtcCode {
  /** OBD-II code e.g. P0300 */
  code: string;
  /** Short human-readable title (Turkish) */
  title: string;
  /** Vehicle system identifier */
  system: string;
  /** Human-readable system name (Turkish) */
  systemLabel?: string;
  severity: DtcCodeSeverity;
  /** Full description of the fault (Turkish) */
  description: string;
  /** Observable symptoms (Turkish) */
  symptoms: string[];
  /** Possible causes (Turkish) */
  causes: string[];
  repairSteps: RepairStep[];
  /** Approximate repair cost range in Turkish Lira */
  estimatedCost: string;
  /** DIY difficulty level */
  difficulty: DtcCodeDifficulty;
  relatedCodes?: string[];
}