# lib/api-zod/src/generated/types/repairStep.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lexus GS300 Arıza Teşhis API
 * OpenAPI spec version: 0.1.0
 */

export interface RepairStep {
  order: number;
  description: string;
  /** @nullable */
  estimatedTime?: string | null;
  /** @nullable */
  toolRequired?: string | null;
}