# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lexus GS300 Arıza Teşhis API
 * OpenAPI spec version: 0.1.0
 */
import type { StatsSummaryBySeverity } from './statsSummaryBySeverity';
import type { StatsSummaryBySystemItem } from './statsSummaryBySystemItem';

export interface StatsSummary {
  totalCodes: number;
  bySeverity: StatsSummaryBySeverity;
  bySystem: StatsSummaryBySystemItem[];
}