# lib/api-zod/src/generated/types/symptomCategory.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lexus GS300 Arıza Teşhis API
 * OpenAPI spec version: 0.1.0
 */
import type { Symptom } from './symptom';

export interface SymptomCategory {
  id: string;
  label: string;
  symptoms: Symptom[];
}