# lib/api-zod/src/generated/types/symptomInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lexus GS300 Arıza Teşhis API
 * OpenAPI spec version: 0.1.0
 */

export interface SymptomInput {
  /** List of symptom IDs or free text descriptions */
  symptoms: string[];
  /**
     * Optional vehicle system to narrow search
     * @nullable
     */
  system?: string | null;
}